shared_requirements.txt
fastapi==0.112.0          # The web framework for building the API
httpx==0.24.1             # HTTP client for making requests (e.g., to other services)
pytest==7.4.0             # Testing framework
pytest-asyncio==0.21.0    # Support for asyncio in pytest
requests==2.32.3          # HTTP library for making requests (popular and simple to use)
pytest-mock==3.14.0       # Mocking library for pytest
python-multipart==0.0.9   # Run services using multithreading
uvicorn==0.30.6           # ASGI server for running FastAPI applications


passlib==1.7.4            # Password hashing library with bcrypt support
bcrypt==4.2.0             # Password hashing library with bcrypt support
pydantic==2.8.2           # Data validation and settings management
pydantic[email]           # Email validation plugin for Pydantic
sqlalchemy==2.0.0         # ORM for database interaction
alembic==1.11.1           # Database migrations tool compatible with SQLAlchemy
#psycopg2==2.9.9          # PostgreSQL database adapter
python-dotenv==1.0.0      # Loading environment variables from a shared .env.production file
psycopg2-binary==2.9.9    # PostgreSQL database adapter
cryptography==43.0.1      # Cryption and decryption of user id